+++ packages/primitives/src/components/Form/index.tsx
@@ -1,7 +1,7 @@
 import Field from '@components/Form/Field';
+import RadioField from '@components/Form/Field/RadioField';
+import TextField from '@components/Form/Field/TextField';
 import FormRoot from '@components/Form/Form';
-import RadioField from '@components/Form/RadioField';
-import TextField from '@components/Form/TextField';
 
 const Form = Object.assign(FormRoot, {
   TextField,
